Create QuizQuestion for question content
